+++ mocks/mocks/mocks.js
@@ -1,0 +1,10 @@
+module.exports = [
+  {
+    id: "base",
+    routesVariants: ["get-user:1"],
+  },
+  {
+    id: "user2",
+    routesVariants: ["get-user:2"],
+  },
+];