+++ test/index.spec.js
@@ -1,17 +1,28 @@
+import { wait } from "./support/helpers";
+
 import {
   readAbout,
+  readSettings,
+  updateSettings,
+  readAlerts,
+  readAlert,
+  readMocks,
+  readMock,
+  readRoutes,
+  readRoute,
+  readRoutesVariants,
+  readRouteVariant,
+  readCustomRoutesVariants,
+  useRouteVariant,
+  restoreRoutesVariants,
+  // legacy
   readBehaviors,
   readBehavior,
   readFixtures,
   readFixture,
-  readSettings,
-  readAlerts,
-  readAlert,
-  updateSettings,
-  wait,
-} from "./support/helpers";
+} from "../index";
 
-describe("react-admin-client methods used through node", () => {
+describe("react-admin-client methods used with node", () => {
   describe("when reading about", () => {
     it("should return current version", async () => {
       const about = await readAbout();
@@ -20,14 +31,26 @@ describe("react-admin-client methods used through node", () => {
   });
 
   describe("when reading alerts", () => {
-    describe("when there are no alerts", () => {
-      it("should return an empty array", async () => {
+    describe("when there are alerts", () => {
+      it("should return alerts", async () => {
         const alerts = await readAlerts();
-        expect(alerts).toEqual([]);
+        expect(alerts.length).toEqual(2);
+      });
+
+      it("should return alert about legacy usage", async () => {
+        const alerts = await readAlerts();
+        expect(alerts[0].message).toEqual(expect.stringContaining("Legacy mocks enabled"));
+      });
+
+      it("should return alert about mock not defined", async () => {
+        const alerts = await readAlerts();
+        expect(alerts[1].message).toEqual(
+          expect.stringContaining('Option "mock" was not defined')
+        );
       });
     });
 
-    describe("when there are alerts", () => {
+    describe("when there are alerts about files with error", () => {
       it("should return alerts array", async () => {
         expect.assertions(1);
         await updateSettings({
@@ -35,25 +58,27 @@ describe("react-admin-client methods used through node", () => {
         });
         await wait(2000);
         const alerts = await readAlerts();
-        expect(alerts.length).toEqual(1);
+        expect(alerts.length).toEqual(4);
       });
 
-      it("first alert model should exist", async () => {
+      it("alert about files error should exist", async () => {
         const alerts = await readAlerts();
-        const alertId = alerts[0].id;
+        const alertId = alerts[3].id;
         const alert = await readAlert(alertId);
         expect(alert.id).toEqual(alertId);
+        expect(alert.message).toEqual(expect.stringContaining("Error loading mocks"));
       });
     });
 
     describe("when alerts are removed", () => {
-      it("should return an empty array again", async () => {
+      it("should return only one alert about legacy usage", async () => {
         await updateSettings({
           path: "mocks",
+          mock: "base",
         });
         await wait(2000);
         const alerts = await readAlerts();
-        expect(alerts).toEqual([]);
+        expect(alerts.length).toEqual(1);
       });
     });
   });
@@ -112,13 +137,6 @@ describe("react-admin-client methods used through node", () => {
       const fixture = await readFixture(fixtureId);
       expect(fixture.id).toEqual(fixtureId);
     });
-
-    it("first fixture of behavior user2 should exist", async () => {
-      const behaviors = await readBehaviors();
-      const fixtureId = behaviors[1].fixtures[0];
-      const fixture = await readFixture(fixtureId);
-      expect(fixture.id).toEqual(fixtureId);
-    });
   });
 
   describe("when reading settings", () => {
@@ -157,6 +175,127 @@ describe("react-admin-client methods used through node", () => {
       });
       const settings = await readSettings();
       expect(settings.behavior).toEqual("base");
+    });
+  });
+
+  describe("when reading mocks", () => {
+    it("should return mocks", async () => {
+      const mocks = await readMocks();
+      expect(mocks).toEqual([
+        {
+          id: "base",
+          from: null,
+          routesVariants: ["get-user:1"],
+          appliedRoutesVariants: ["get-user:1"],
+        },
+        {
+          id: "user2",
+          from: null,
+          routesVariants: ["get-user:2"],
+          appliedRoutesVariants: ["get-user:2"],
+        },
+      ]);
+    });
+  });
+
+  describe("when reading mock", () => {
+    it("should return mock data", async () => {
+      const mock = await readMock("base");
+      expect(mock).toEqual({
+        id: "base",
+        from: null,
+        routesVariants: ["get-user:1"],
+        appliedRoutesVariants: ["get-user:1"],
+      });
+    });
+  });
+
+  describe("when reading routes", () => {
+    it("should return routes", async () => {
+      const data = await readRoutes();
+      expect(data).toEqual([
+        {
+          id: "get-user",
+          delay: null,
+          url: "/api/user",
+          method: "GET",
+          variants: ["get-user:1", "get-user:2"],
+        },
+      ]);
+    });
+  });
+
+  describe("when reading route", () => {
+    it("should return route data", async () => {
+      const mock = await readRoute("get-user");
+      expect(mock).toEqual({
+        id: "get-user",
+        delay: null,
+        url: "/api/user",
+        method: "GET",
+        variants: ["get-user:1", "get-user:2"],
+      });
+    });
+  });
+
+  describe("when reading routes variants", () => {
+    it("should return routes variants", async () => {
+      const data = await readRoutesVariants();
+      expect(data).toEqual([
+        {
+          id: "get-user:1",
+          routeId: "get-user",
+          handler: "default",
+          response: { body: [{ email: "foo@foo.com" }], status: 200 },
+          delay: null,
+        },
+        {
+          id: "get-user:2",
+          routeId: "get-user",
+          handler: "default",
+          response: { body: [{ email: "foo2@foo2.com" }], status: 200 },
+          delay: null,
+        },
+      ]);
+    });
+  });
+
+  describe("when reading route variant", () => {
+    it("should return route variant data", async () => {
+      const data = await readRouteVariant("get-user:2");
+      expect(data).toEqual({
+        id: "get-user:2",
+        routeId: "get-user",
+        handler: "default",
+        response: { body: [{ email: "foo2@foo2.com" }], status: 200 },
+        delay: null,
+      });
+    });
+  });
+
+  describe("mock custom route variants", () => {
+    it("should be empty", async () => {
+      const data = await readCustomRoutesVariants();
+      expect(data).toEqual([]);
+    });
+
+    it("should be able to add one", async () => {
+      await useRouteVariant("get-user:2");
+      const data = await readCustomRoutesVariants();
+      expect(data).toEqual(["get-user:2"]);
+    });
+
+    it("should reject if route variant don't exist", async () => {
+      expect.assertions(1);
+      await useRouteVariant("foo").catch((err) => {
+        expect(err.message).toEqual('Route variant with id "foo" was not found');
+      });
+    });
+
+    it("should be empty after restoring them to the mock defaults", async () => {
+      await restoreRoutesVariants();
+      const data = await readCustomRoutesVariants();
+      expect(data).toEqual([]);
     });
   });
 });